+++ src/sync_ftp/mod.rs
@@ -464,13 +464,14 @@ impl FtpStream {
     /// ### put_file
     ///
     /// This stores a file on the server.
-    /// r argument must be any struct which implemenents the Read trait
-    pub fn put_file<R: Read>(&mut self, filename: &str, r: &mut R) -> Result<()> {
+    /// r argument must be any struct which implemenents the Read trait.
+    /// Returns amount of written bytes
+    pub fn put_file<R: Read>(&mut self, filename: &str, r: &mut R) -> Result<u64> {
         // Get stream
         let mut data_stream = self.put_with_stream(filename)?;
-        copy(r, &mut data_stream)
-            .map_err(FtpError::ConnectionError)
-            .and_then(|_| self.finalize_put_stream(Box::new(data_stream)))
+        let bytes = copy(r, &mut data_stream).map_err(FtpError::ConnectionError)?;
+        self.finalize_put_stream(Box::new(data_stream))?;
+        Ok(bytes)
     }
 
     /// ### put_with_stream